% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{age_plus_one_plot}
\alias{age_plus_one_plot}
\title{Function to extract  population sizes of age +1 individuals}
\usage{
age_plus_one_plot(data)
}
\arguments{
\item{dataframe}{summary_classAllTime.csv dataframe}
}
\value{
ggplot
}
\description{
This function produces a  plot with population sizes of age 1+ individuals. It is based on the N_Initial_class column of the CDMetaPOP output file summary_classAllTime.
}
