% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diagnostics.R
\name{alleles_by_year}
\alias{alleles_by_year}
\title{Function extracts and plots allele count overtime}
\usage{
alleles_by_year(data, n = 5)
}
\arguments{
\item{n}{An integer that specifies the time intervals of the time series. It defaults to 5 years.}

\item{dataframe}{summary_popAllTime.csv}
}
\value{
a ggplot object representing the overtime observed and expected heterozygosities values in the simulation.
}
\description{
This function plots the total number of unique alleles for every year of the simulation. It is based on the Allele column of the CDMetaPOP output file summary_popAllTime
}
