% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{births}
\alias{births}
\title{Create a plot with the proportion of young of the year. Note here that this proportion includes only age 0 progeny that survived egg mortality.}
\usage{
births(file_path)
}
\arguments{
\item{x}{Specify the path to the summary_popAllTime.csv dataframe}
}
\value{
a ggplot object representing the size of progeny surviving egg deaths every year
}
\description{
This function plots the proportion of progeny over the total, after egg deaths are taken into account
}
\examples{
path <- system.file("extdata", "summary_popAllTime.csv", package = "cdmetapopR")
births(path)
}
