% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmetapop_to_gene.R
\name{cdmetapop_to_gene}
\alias{cdmetapop_to_gene}
\title{Convert CDMetaPOP data to GENEPOP or GENALEX format}
\usage{
cdmetapop_to_gene(path, format = "genepop")
}
\arguments{
\item{path}{Character string; the file path to the \code{CDMetaPOP} CSV file.}

\item{format}{A character string indicating the desired output format: "genepop" or "genalex". Genepop is default.}
}
\value{
The function writes the GENEPOP or GENALEX formatted data into the working directory.
}
\description{
This function reads a \code{CDMetaPOP} CSV file, processes the genetic loci, and converts the data into \code{GENEPOP} or \code{GENALEX} formats.
}
\examples{
\dontrun{
# Convert a CDMetaPOP file to GENEPOP format
cdmetapop_to_gene(
  path = "your_path/ind1.csv",
  format = "genepop"
)

# Convert a CDMetaPOP file to GENALEX format
cdmetapop_to_gene(
  path = "your_path/ind1.csv",
  format = "genalex"
)
}
}
