% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmetapop_to_genepop.R
\name{cdmetapop_to_genepop}
\alias{cdmetapop_to_genepop}
\title{Convert CDMetaPOP data to GENEPOP format}
\usage{
cdmetapop_to_genepop(path, stratum, output_file)
}
\arguments{
\item{path}{Character string; the file path to the \code{CDMetaPOP} CSV file.}

\item{stratum}{Character string; the column name in the input data that specifies the population (e.g., \code{PatchID}).}

\item{output_file}{Character string; the file path for the output GENEPOP file.}
}
\value{
The function writes the GENEPOP-formatted data to the specified file.
}
\description{
This function reads a \code{CDMetaPOP} CSV file, processes the genetic loci, and converts the data into \code{GENEPOP} format.
}
\examples{
\dontrun{
# Convert a CDMetaPOP file to GENEPOP format
cdmetapop_to_genepop(
  path = "your_path/ind1.csv",
  stratum = "PatchID",
  output_file = "cdmetapop_to_genepop_output.gen"
)
}
}
