% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cdmat.R
\name{create_cdmat}
\alias{create_cdmat}
\title{Create cost distance matrix for CDMetaPOP simulation}
\usage{
create_cdmat(
  coords,
  method = c("euclidean", "equal", "lcp"),
  resistance = NULL
)
}
\arguments{
\item{coords}{A set of spatial coordinates in the form of a 2-column data frame or matrix}

\item{method}{Method for creating the cost distance matrix, must be "euclidean", "equal", or "lcp"}

\item{resistance}{Raster object representing resistance surface for generating cost distance matrix}
}
\value{
An NxN matrix of cost distances among patch locations
}
\description{
Takes input coordinates and produces a symmetrical cost distance matrix with options for Euclidean distance, equal distance, or least cost paths based on a provided resistance surface
}
\details{
Hello
}
\examples{
x <- c(1,2,3,4,5,6)
y <- c(1,2,3,4,5,6)
test <- data.frame(x=x, y=y)
create_cdmat(coords=test)

x <- runif(6, min=0, max=10)
y <- runif(6, min=0, max=10)
coords <- data.frame(x=x, y=y)
r <- terra::rast(nrows=10, ncols=10, xmin = 0, xmax = 10, ymin = 0, ymax = 10)
terra::values(r) <- 1
create_cdmat(coords=coords, method="lcp", resistance=r)
}
