% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Counts number of individuals that move each year by reading in each ind file}
\usage{
dispersal(
  path = path,
  run = 0,
  batch = 0,
  mc = 0,
  gen = 49,
  species = 0,
  ind = TRUE,
  plot = FALSE
)
}
\arguments{
\item{path}{Character. The path to the directory containing the output files.}

\item{run}{Integer. The run index. A run is each row of the RunVars file - Defaults to 0.}

\item{batch}{Integer. The batch index. See RunVars PreProcess Run parameters in the CDMetaPOP user manual.}

\item{mc}{Integer. The Monte Carlo index. It refers to the number of Monte Carlo simulations output files to examine. It refers to the mcruns option in the RunVars CDMetaPOP input file. It defaults to 0, which is one run.}

\item{gen}{Integer. The number of generations. Simulation run time (generation or year). Refers to runtime in the RunVars of CDMetaPOP input file. It defaults to 49.}

\item{species}{Integer. The species index. If the simulation includes multiple species. It defaults to 0, meaning one species only.}

\item{ind}{Logical. Whether to use ind files ('TRUE') or sample files ('FALSE').}

\item{plot}{Logical. If TRUE, a histogram of the proportion of movers across the ind file defined by the chosen parameters will be displayed. Defaults to FALSE. NOTE: The histogram of movers only reflects the counts for the given combination of parameters (not across all iterations).}
}
\value{
A data frame with proportions of movers or a plot of the count of movers per year. The data returned is a dataframe that lists the proportion of dispersors for each year and for each Monte Carlo simulation.
}
\description{
Counts number of individuals that move each year by reading in each ind file
}
\details{
This function was partially developed with assistance from ChatGPT, an AI model by OpenAI.
}
\examples{
\dontrun{ 
# Example: This function is typically used with large datasets.See the package vignette for a full example.
# Run the function to output a dataframe for all the generations and replicas specified:
dispersal (path = my_dataset, run = 1, batch = 0, mc = 2, gen = 9, species = 0, ind = FALSE, plot = FALSE)
}
\dontrun{ 
# Example: This function is typically used with large datasets.See the package vignette for a full example.
# Generating a barplot for the example directory
dispersal(path = my_dataset, run = 0, batch = 0, mc = 0, gen = 9, species = 1, ind = FALSE, plot = TRUE)
}
}
