% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{extinct_patch}
\alias{extinct_patch}
\title{Check which patch(es) went extinct (no more individuals left) at a particular year}
\usage{
extinct_patch(data, year = 100)
}
\arguments{
\item{year}{value referring to the year to select. It defaults to 100}

\item{dataframe}{summary_popAllTime.csv dataframe}
}
\value{
integer vector
}
\description{
This function produces a vector with the number of patches that suffered extinction at a given year. It is based off the N_Initial column of the CDMetaPOP output file summary_popAllTime.
}
\examples{
path <- system.file("extdata", "summary_popAllTime.csv", package = "cdmetapopR")
data <- read.csv(path)
extinct_patch(data, year = 8)
}
