% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{locus}
\alias{locus}
\title{Deprecated 'locus' function from gstudio}
\source{
The original implementation is from the \code{gstudio} package
(\url{https://github.com/dyerlab/gstudio}).
}
\usage{
locus(x, type = "codom", phased = FALSE)
}
\description{
This function is a copy of the deprecated locus function from the \code{gstudio} package.
It has been included in this package because \code{gstudio} is no longer maintained.
}
\note{
This function is provided as-is, without guarantees of future updates.
Users are encouraged to transition to more modern alternatives if available.
}
\section{Deprecated}{

This function is deprecated and may be removed in future versions of this package.
Users will see a warning message if this function is called.
}

\examples{
\dontrun{
  # Example usage
  result <- locus(c("0", "1"), type = "snp")
}

}
\keyword{internal}
