% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diseasevars.R
\name{make_diseasevars}
\alias{make_diseasevars}
\title{Build DiseaseVars File from Template}
\usage{
make_diseasevars(output_file = "my_new_diseasevars.csv")
}
\arguments{
\item{output_file}{The name of the output file. Defaults to 'my_new_diseasevars.csv'.}
}
\value{
A Shiny app instance.
}
\description{
This function loads a provided DiseaseVars template, allows the user to edit selected areas,
and saves the modified version as a new file.
}
