% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_patchvars.R
\name{make_patchvars}
\alias{make_patchvars}
\title{Build PatchVars File from Template}
\usage{
make_patchvars(output_file = "my_new_patchvars.csv")
}
\arguments{
\item{output_file}{The name of the output file. Defaults to 'my_new_patchvars.csv'.}
}
\value{
A Shiny app instance.
}
\description{
This function loads a provided PatchVars template, allows the user to edit selected areas,
and saves the modified version as a new file.
}
