% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_runvars.R
\name{make_runvars}
\alias{make_runvars}
\title{Build RunVars File from Template}
\usage{
make_runvars(output_file = "my_new_runvars.csv")
}
\arguments{
\item{output_file}{The name of the output file. Defaults to 'my_new_runvars.csv'.}
}
\value{
A Shiny app instance.
}
\description{
This function loads a provided runVars template, allows the user to edit selected areas,
and saves the modified version as a new file.
}
