% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{pop_count_plot}
\alias{pop_count_plot}
\title{Create abundance by year plot}
\usage{
pop_count_plot(data)
}
\arguments{
\item{dataframe}{Import the summary_popAllTime.csv dataframe}
}
\value{
a ggplot object representing the overtime general population size
}
\description{
This function plots population dynamics overtime showing a count of the individuals over the years of the simulation. It is based off the N_Initial column of the CDMetaPOP output file summary_popAllTime
}
\examples{
path <- system.file("extdata", "summary_popAllTime.csv", package = "cdmetapopR")
data <- read.csv(path)
pop_count_plot(data)
}
