% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{pop_patch_count}
\alias{pop_patch_count}
\title{Create abundance by patch plot}
\usage{
pop_patch_count(data, years = c(1, 25, 50, 100))
}
\arguments{
\item{years}{vector of values referring to the years to select. It defaults to c(1,25,50,100)}

\item{dataframe}{summary_popAllTime.csv dataframe}
}
\value{
a ggplot boxplot showing the population numbers of different patches
}
\description{
This function plots the population abundance per Patch. It is based off the N_Initial column of the CDMetaPOP output file summary_popAllTime.
}
