% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{pop_sex_plot}
\alias{pop_sex_plot}
\title{Create abundance by year plots for the different sex - Male/Female/MYY/FYY}
\usage{
pop_sex_plot(data)
}
\arguments{
\item{dataframe}{Import the summary_popAllTime.csv dataframe}
}
\value{
a ggplot object representing the overtime  population size for the four sexes
}
\description{
This function plots population dynamics overtime showing the count of the individuals separated by sex over the years of the simulation. It is based off the following columns: N_Males_1, N_YYMales_1, N_Females_1 and N_YYFemales of the CDMetaPOP output file summary_popAllTime
}
\examples{
path <- system.file("extdata", "summary_popAllTime.csv", package = "cdmetapopR")
data <- read.csv(path)
pop_sex_plot(data)
}
