% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdmetapop.R
\name{read.cdmetapop}
\alias{read.cdmetapop}
\title{Read cdmetapop file as dataframe}
\usage{
read.cdmetapop(file_path)
}
\arguments{
\item{file_path}{Specify the file name and path.}
}
\value{
Dataframe
}
\description{
This function imports in r a cdmetapop output file as a dataframe
}
