% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size_age_class}
\alias{size_age_class}
\title{Function to create a plot of size at age overtime}
\usage{
size_age_class(file_path)
}
\arguments{
\item{x}{Specify the path to the summary_popAllTime.csv dataframe}
}
\value{
a ggplot object representing the overtime average sizes of the individuals in the simulation separated by age.
}
\description{
This function plots the overtime average size of individuals at a specific age. It is based on the AgeSize_mean columns of the CDMetaPOP output file summary_classAllTime
}
