% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_conversions.R
\name{unite_column}
\alias{unite_column}
\title{Unite multiple columns into one}
\usage{
unite_column(dataframe, column_name, sep, cols)
}
\arguments{
\item{dataframe}{A data frame}

\item{column_name}{Name of the new column}

\item{sep}{Separator for united columns}

\item{cols}{Columns to unite (character vector of column names)}
}
\value{
A data frame with the united columns
}
\description{
A wrapper for tidyr::unite designed specifically to handle large cdmetapop column outputs
}
\examples{
# example code
df <- data.frame(
  Subpatch = c("A", "B", "C"),
  N = c(25, 30, 35),
  N_1 = c(34, 44, 34),
  N_2 = c(21, 43, 54)
)

# Specify the columns to unite
cols_to_unite <- c("N_1", "N_2")

# Call the function to unite the columns
result <- unite_column(df, column_name = "N", sep = "|", cols = cols_to_unite)
print(result)

# Example with column names X1 to X100
df2 <- data.frame(matrix(1:10000, ncol = 100))
cols <- names(df2)
result2 <- unite_column(df2, column_name = "All_X", sep = ",", cols = cols)
head(result2)


}
