% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_classvars.R
\name{write_classvars}
\alias{write_classvars}
\title{Build ClassVars File from Template}
\usage{
write_classvars(output_file = "my_new_classvars.csv")
}
\arguments{
\item{output_file}{The name of the output file. Defaults to 'my_new_classvars.csv'.}
}
\value{
A Shiny app instance.
}
\description{
This function loads a provided ClassVars template, allows the user to edit selected areas,
and saves the modified version as a new file.
}
