% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_popvars.R
\name{write_popvars}
\alias{write_popvars}
\title{Build PopVars File from Template}
\usage{
write_popvars(output_file = "my_new_popvars.csv")
}
\arguments{
\item{output_file}{The name of the output file. Defaults to 'my_new_popvars.csv'.}
}
\value{
A Shiny app instance.
}
\description{
This function loads a provided PopVars template, allows the user to edit selected areas,
and saves the modified version as a new file.
}
