% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_cdmetapop.R
\name{launch_cdmetapop}
\alias{launch_cdmetapop}
\title{Function to facilitate the running of CDMetaPOP software from R}
\usage{
launch_cdmetapop(
  pythonFilepath = "python",
  CDMetaPOPFilepath = "CDMetaPOP.py",
  runvarsDirectory = NULL,
  runvarsFilename = "RunVars.csv",
  outputDirectory = "test_"
)
}
\arguments{
\item{pythonFilepath}{Location of Python executable, or just 'python' if the environment is already established}

\item{CDMetaPOPFilepath}{Location of the CDMetaPOP.py file, if already in the working directory then this can just be 'CDMetaPOP.py'}

\item{runvarsDirectory}{Location of the directory where the RunVars.csv file is stored for the run}

\item{runvarsFilename}{Name of the RunVars file to run}

\item{outputDirectory}{Name of output directory to be stored in the same location as runvarsDirectory}
}
\description{
Function to run CDMetaPOP from R
}
\details{
This function will launch a command prompt (Windows) or terminal (Linux/Mac) that will call Python and supply the 5 arguments needed to launch CDMetaPOP simulations
}
\examples{
\dontrun{ 
# Example: This function launches a system command to initiate CDMetaPOP.
pythonFilepath="C:/Users/User1/anaconda3/python.exe" 
CDMetaPOPFilepath="C:/Users/User1/CDMetaPOP_v3.03/src/CDMetaPOP.py"
runvarsDirectory = "C:/Users/User1/CDMetaPOP_v3.03/example_files/"
runvarsFilename = "RunVars.csv"
outputDirectory = "test_" 

mymodel <- launch_cdmetapop(pythonFilepath,
                           CDMetaPOPFilepath,
                           runvarsDirectory,
                           runvarsFilename,
                           outputDirectory = "test_")
}
}
