% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdmetapop.R
\name{read_cdmetapop}
\alias{read_cdmetapop}
\title{Read an existing CDMetaPOP input file into an R6 object}
\usage{
read_cdmetapop(filepath, type = "ClassVars")
}
\arguments{
\item{filepath}{Path to an existing CDMetaPOP input csv.}

\item{type}{Which input file type \code{filepath} is. Currently \code{"ClassVars"}
and \code{"PatchVars"} are supported. Case-insensitive.}
}
\value{
An R6 object of the corresponding class (e.g. a \code{\link[=ClassVars]{ClassVars()}}
object), with \code{location} set to \code{filepath}.
}
\description{
Read an existing CDMetaPOP input file into an R6 object
}
\examples{
\dontrun{
myclassvars <- read_cdmetapop("classvars/ClassVars_AS1.csv", type = "ClassVars")
mypatchvars <- read_cdmetapop("patchvars/PatchVarsS1.csv", type = "PatchVars")
}
}
